Change | logic improvement in the socials part. Styling added to sign up form
