+++ src/components/pages/Signup/style.css
@@ -1,0 +1,3 @@
+.form-height {
+    min-height: 700px;
+}
